% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_all.R
\name{deploy_apps}
\alias{deploy_apps}
\title{Deploy apps to a server}
\usage{
deploy_apps(account = "barret", server = "shinyapps.io", apps = TRUE,
  libpath = "_shiny-examples-lib", cores = 3)
}
\arguments{
\item{account, server}{args supplied to \code{[rsconnect::deployApp]}}

\item{apps}{A vector of three digit character values or \code{TRUE} to deploy all apps}

\item{libpath}{library location. (Creates the path if it does not exist.)}

\item{cores}{number of cores to use when deploying}
}
\description{
Deploy apps to a server
}
