% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_public.R
\name{set_public}
\alias{set_public}
\title{Set applications to be public}
\usage{
set_public(account, server, api_key)
}
\arguments{
\item{account, server}{args supplied to \code{[rsconnect::deployApp]}}

\item{api_key}{API key generated from the connect server}
}
\description{
Set applications to be public
}
\examples{
\dontrun{
  set_public("barret", "beta.rstudioconnect.com", "APIKEY")
}
}
